+++ integration-test/src/test/java/org/jboss/pnc/integration/ProcessProgressNotificationTest.java
@@ -20,15 +20,17 @@ package org.jboss.pnc.integration;
 import org.jboss.arquillian.container.test.api.Deployment;
 import org.jboss.arquillian.junit.Arquillian;
 import org.jboss.pnc.AbstractTest;
+import org.jboss.pnc.common.json.JsonOutputConverterMapper;
+import org.jboss.pnc.dto.Build;
+import org.jboss.pnc.enums.BuildCoordinationStatus;
 import org.jboss.pnc.integration.deployments.Deployments;
 import org.jboss.pnc.integration.websockets.NotificationCollector;
+import org.jboss.pnc.mock.dto.BuildMock;
 import org.jboss.pnc.rest.notifications.websockets.Action;
 import org.jboss.pnc.rest.notifications.websockets.MessageType;
 import org.jboss.pnc.rest.notifications.websockets.NotificationsEndpoint;
 import org.jboss.pnc.rest.notifications.websockets.ProgressUpdatesRequest;
 import org.jboss.pnc.rest.notifications.websockets.TypedMessage;
-import org.jboss.pnc.common.json.JsonOutputConverterMapper;
-import org.jboss.pnc.enums.BuildCoordinationStatus;
 import org.jboss.pnc.spi.coordinator.events.DefaultBuildStatusChangedEvent;
 import org.jboss.pnc.spi.events.BuildCoordinationStatusChangedEvent;
 import org.jboss.pnc.spi.events.BuildSetStatusChangedEvent;
@@ -49,10 +51,8 @@ import javax.websocket.ContainerProvider;
 import javax.websocket.RemoteEndpoint;
 import javax.websocket.Session;
 import javax.websocket.WebSocketContainer;
-
 import java.lang.invoke.MethodHandles;
 import java.net.URI;
-import java.util.Date;
 import java.util.function.Supplier;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -82,6 +82,7 @@ public class ProcessProgressNotificationTest {
         WebArchive restWar = enterpriseArchive.getAsType(WebArchive.class, AbstractTest.REST_WAR_PATH);
         restWar.addClass(ProcessProgressNotificationTest.class);
         restWar.addClass(NotificationCollector.class);
+        restWar.addPackages(true, BuildMock.class.getPackage());
         restWar.addPackage(NotificationsEndpoint.class.getPackage());
         restWar.addPackage(Notifier.class.getPackage());
         logger.info(enterpriseArchive.toString(true));
@@ -104,8 +105,13 @@ public class ProcessProgressNotificationTest {
 
         // given
         Integer taskId = 1;
-        BuildCoordinationStatusChangedEvent buildStatusChangedEvent = new DefaultBuildStatusChangedEvent(BuildCoordinationStatus.NEW,
-                BuildCoordinationStatus.DONE, taskId, 1, 1, "Build1", new Date(1453118400000L), new Date(1453122000000L), 1);
+
+        Build build = BuildMock.newBuild(taskId, BuildCoordinationStatus.DONE, "Build1");
+
+        BuildCoordinationStatusChangedEvent buildStatusChangedEvent = new DefaultBuildStatusChangedEvent(
+                build,
+                BuildCoordinationStatus.NEW
+        );
 
         //when
         buildStatusNotificationEvent.fire(buildStatusChangedEvent);
@@ -119,8 +125,9 @@ public class ProcessProgressNotificationTest {
         waitForMessages(1);
 
         //then
-        logger.info("Received: " + notificationCollector.getMessages().get(0)); //"eventType":"BUILD_STATUS_CHANGED","payload":{"id":1,"buildCoordinationStatus":"DONE
-        assertThat(notificationCollector.getMessages().get(0)).startsWith("{\"eventType\":\"BUILD_STATUS_CHANGED\",\"payload\":{\"id\":1,\"buildCoordinationStatus\":\"DONE\"");
+        logger.info("Received: " + notificationCollector.getMessages().get(0));
+
+        assertThat(notificationCollector.getMessages().get(0)).startsWith("{\"eventType\":\"BUILD_STATUS_CHANGED\",\"payload\":{\"oldStatus\":\"NEW\",\"build\":");
     }
 
     private void waitForMessages(int numberOfMessages) {