+++ integration-test/src/test/java/org/jboss/pnc/integration/remote/RemoteBuildTest.java
@@ -80,8 +80,8 @@ public class RemoteBuildTest {
 
                 if (EventType.BUILD_STATUS_CHANGED.equals(notification.getEventType())) {
                     BuildChangedPayload buildStatusUpdate = (BuildChangedPayload) notification.getPayload();
-                    if (buildStatusUpdate.getBuildCoordinationStatus().isCompleted()) {
-                        notifyCompleted(buildStatusUpdate.getBuildConfigurationId(), buildStatusUpdate.getBuildCoordinationStatus());
+                    if (buildStatusUpdate.getBuild().getStatus().isCompleted()) {
+                        notifyCompleted(buildStatusUpdate.getBuild().getBuildConfigurationAudited().getId(), buildStatusUpdate.getBuild().getStatus());
                     }
                 }
             } catch (IOException e) {