+++ pnc-mock/src/main/java/org/jboss/pnc/mock/dto/SCMRepositoryMock.java
@@ -15,15 +15,21 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jboss.pnc.spi.events;
+package org.jboss.pnc.mock.dto;
 
-import org.jboss.pnc.dto.Build;
-import org.jboss.pnc.enums.BuildCoordinationStatus;
+import org.jboss.pnc.dto.SCMRepository;
 
-public interface BuildCoordinationStatusChangedEvent {
+/**
+ * @author <a href="mailto:matejonnet@gmail.com">Matej Lazar</a>
+ */
+public class SCMRepositoryMock {
 
-    BuildCoordinationStatus getOldStatus();
-
-    Build getBuild();
-
+    public static SCMRepository newScmRepository() {
+        return SCMRepository.builder()
+            .id(1)
+            .internalUrl("url1")
+            .externalUrl("url2")
+            .preBuildSyncEnabled(true)
+            .build();
+    }
 }