+++ spi/src/main/java/org/jboss/pnc/spi/coordinator/events/DefaultBuildStatusChangedEvent.java
@@ -17,67 +17,20 @@
  */
 package org.jboss.pnc.spi.coordinator.events;
 
+import org.jboss.pnc.dto.Build;
 import org.jboss.pnc.enums.BuildCoordinationStatus;
 import org.jboss.pnc.spi.events.BuildCoordinationStatusChangedEvent;
-
-import java.util.Date;
 
 public class DefaultBuildStatusChangedEvent implements BuildCoordinationStatusChangedEvent {
 
     private final BuildCoordinationStatus oldStatus;
-    private final BuildCoordinationStatus newStatus;
-    private final Integer buildTaskId;
-    private final Integer userId;
-    private final Integer buildConfigurationId;
-    private final Integer buildConfigurationRevision;
-    private final String buildConfigurationName;
-    private final Date buildStartTime;
-    private final Date buildEndTime;
+    private final Build build;
 
     public DefaultBuildStatusChangedEvent(
-            BuildCoordinationStatus oldStatus,
-            BuildCoordinationStatus newStatus,
-            Integer buildTaskId,
-            Integer buildConfigurationId,
-            Integer buildConfigurationRevision,
-            String buildConfigurationName,
-            Date buildStartTime,
-            Date buildEndTime,
-            Integer userId) {
+            Build build,
+            BuildCoordinationStatus oldStatus) {
+        this.build = build;
         this.oldStatus = oldStatus;
-        this.newStatus = newStatus;
-        this.buildTaskId = buildTaskId;
-        this.buildConfigurationId = buildConfigurationId;
-        this.buildConfigurationRevision = buildConfigurationRevision;
-        this.buildConfigurationName = buildConfigurationName;
-        this.buildStartTime = buildStartTime;
-        this.buildEndTime = buildEndTime;
-        this.userId = userId;
-    }
-
-    @Override
-    public Integer getBuildTaskId() {
-        return buildTaskId;
-    }
-
-    @Override
-    public Integer getUserId() {
-        return userId;
-    }
-
-    @Override
-    public Integer getBuildConfigurationId() {
-        return buildConfigurationId;
-    }
-
-    @Override
-    public Integer getBuildConfigurationRevision() {
-        return buildConfigurationRevision;
-    }
-
-    @Override
-    public String getBuildConfigurationName() {
-        return buildConfigurationName;
     }
 
     @Override
@@ -86,18 +39,8 @@ public class DefaultBuildStatusChangedEvent implements BuildCoordinationStatusCh
     }
 
     @Override
-    public BuildCoordinationStatus getNewStatus() {
-        return newStatus;
-    }
-
-    @Override
-    public Date getBuildStartTime() {
-        return buildStartTime;
-    }
-
-    @Override
-    public Date getBuildEndTime() {
-        return buildEndTime;
+    public Build getBuild() {
+        return build;
     }
 
     @Override
@@ -105,13 +48,7 @@ public class DefaultBuildStatusChangedEvent implements BuildCoordinationStatusCh
 
         return "DefaultBuildStatusChangedEvent{" +
                 "oldStatus=" + oldStatus +
-                ", newStatus=" + newStatus +
-                ", buildTaskId=" + buildTaskId +
-                ", userId=" + userId +
-                ", buildConfigurationId=" + buildConfigurationId +
-                ", buildConfigurationName=" + buildConfigurationName +
-                ", buildStartTime=" + buildStartTime +
-                ", buildEndTime=" + buildEndTime +
+                "build=" + build +
                 '}';
     }
 }